+++ __tests__/hooks/useEtherspotHistory.test.js
@@ -1,12 +1,16 @@
 import { renderHook, waitFor } from '@testing-library/react';
+import { ethers } from 'ethers';
 
 // hooks
 import { useEtherspotHistory, EtherspotTransactionKit } from '../../src';
 
+const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'goerli'); // replace with your node's RPC URL
+const provider = new ethers.Wallet.createRandom().connect(ethersProvider);
+
 describe('useEtherspotHistory()', () => {
   it('getAccountTransactions() returns account history', async () => {
     const wrapper = ({ children }) => (
-      <EtherspotTransactionKit provider={null}>
+      <EtherspotTransactionKit provider={provider}>
         {children}
       </EtherspotTransactionKit>
     );
@@ -40,7 +44,7 @@ describe('useEtherspotHistory()', () => {
   });
   it('getAccountTransaction() returns transaction by existing hash', async () => {
     const wrapper = ({ children }) => (
-      <EtherspotTransactionKit provider={null}>
+      <EtherspotTransactionKit provider={provider}>
         {children}
       </EtherspotTransactionKit>
     );